requirements.txt
fastapi
uvicorn
langchain
langchainhub
langchain-openai
langchain-community 
langchain-core
langchain-text-splitters
langdetect
langsmith
openai
twilio
wikipedia
python-decouple
psycopg2-binary
python-multipart
pyngrok
SQLAlchemy
chroma-hnswlib
chromadb
huggingface-hub
opencv-python
pandas
pdf2image
pdfminer.six
pdfplumber
pikepdf
pillow
pillow_heif
pyarrow
pydantic
pypdf
pypdf2
pypdfium2
pytesseract
python-dotenv
PyYAML
rank-bm25
regex
requests
rich
safetensors
scikit-learn
scipy
sentence-transformers
sympy
tiktoken
tokenizers
toml
torch
torchvision
tqdm
transformers
unstructured
unstructured-client
unstructured-inference
unstructured.pytesseract
watchdog
watchfiles
python-decouple
